requirements.txt
tensorflow
streamlit
numpy
pillow
opencv-python
matplotlib
